requirements.txt
fastapi==0.121.0
google-api-python-client==2.186.0
google-auth==2.23.4
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
pyyaml>=6.0
pytest>=7.0.0
ruff>=0.1.0
uvicorn[standard]==0.38.0
